% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveit.R
\name{authMoveIt}
\alias{authMoveIt}
\title{Authenticate User}
\usage{
authMoveIt(baseUrl, payload)
}
\arguments{
\item{baseUrl}{Base URL for Move It server}

\item{payload}{Authentication header}
}
\value{
List of auth tokens for MoveIt access
}
\description{
Authenticate User
}
\examples{
\dontrun{
tokens <- authMoveIt("someurl.com", "auth=grant_type=password&username=USERNAME&password=PASSWORD")
}
}
