% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveit.R
\name{availableFiles}
\alias{availableFiles}
\title{Available Files}
\usage{
availableFiles(baseUrl, tokens)
}
\arguments{
\item{baseUrl}{Base URL for Move It server}

\item{tokens}{List of auth tokens for MoveIt access}
}
\value{
Dataframe of available files and their details
}
\description{
Available Files
}
\examples{
\dontrun{
files <- availableFiles("someurl.com", tokens)
}
}
