% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveit.R
\name{availableFolders}
\alias{availableFolders}
\title{List Available Folder}
\usage{
availableFolders(baseUrl, tokens)
}
\arguments{
\item{baseUrl}{Base URL for Move It server}

\item{tokens}{List of auth tokens for MoveIt access}
}
\value{
Dataframe of available folders and their details
}
\description{
List Available Folder
}
\examples{
\dontrun{
folders <- availableFiles("someurl.com", tokens)
}
}
