% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveit.R
\name{readMoveItFile}
\alias{readMoveItFile}
\title{Function to download a file from the FTP server}
\usage{
readMoveItFile(baseUrl, tokens, id, fileType = "csv")
}
\arguments{
\item{baseUrl}{Base URL for Move It server}

\item{tokens}{List of auth tokens for MoveIt access}

\item{id}{file id}

\item{fileType}{csv or xlsx}
}
\value{
A dataframe of the identified file
}
\description{
Function to download a file from the FTP server
}
\examples{
\dontrun{
df <- readMoveItFile("someurl.com", 626235843, tokens, "csv")
}
}
