% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveit.R
\name{uploadMoveItFile}
\alias{uploadMoveItFile}
\title{Upload file to move it server}
\usage{
uploadMoveItFile(baseUrl, tokens, folderId, filePath, fileType)
}
\arguments{
\item{baseUrl}{Base URL for Move It server}

\item{tokens}{List of auth tokens for MoveIt access}

\item{folderId}{ID for destination folder}

\item{filePath}{Path to file for upload}

\item{fileType}{POST file type ie "text/csv"}
}
\value{
POST content
}
\description{
Upload file to move it server
}
\examples{
\dontrun{
uploadMoveItFile("someurl.com", tokens, 2346247, "some.csv", "text/csv")
}
}
